% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_03_segment_fields_pseudo_units.R
\name{make_fishnet}
\alias{make_fishnet}
\title{Create a fishnet grid over a polygon}
\usage{
make_fishnet(poly_sf, tile_size_m)
}
\arguments{
\item{poly_sf}{sf polygon(s)}

\item{tile_size_m}{numeric tile side length in meters}
}
\value{
sf with square tiles covering the bounding box of poly_sf
}
\description{
Create a fishnet grid over a polygon
}
