% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_00_globals.R
\name{norm_county}
\alias{norm_county}
\title{Normalize county names}
\usage{
norm_county(x)
}
\arguments{
\item{x}{Character vector of county names.}
}
\value{
Character vector of normalized county names.
}
\description{
Utility to normalize county names so that joins between NASS tables,
shapefiles, and validation data are consistent.
}
\details{
Typical usage is to lower case, trim whitespace, and collapse
punctuation / multiple spaces.
}
