% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_03_segment_fields_pseudo_units.R
\name{summarize_tile}
\alias{summarize_tile}
\title{Summarize weight and area inside a tile polygon}
\usage{
summarize_tile(tile_sf, w_smooth, cell_ha)
}
\arguments{
\item{tile_sf}{1-row sf polygon}

\item{w_smooth}{SpatRaster of weights (same CRS as tile_sf)}

\item{cell_ha}{SpatRaster of cell areas in ha}
}
\value{
list(area_ha, w_sum) or NULL if tile area <= 0
}
\description{
Summarize weight and area inside a tile polygon
}
